Add base move-in-out-settings service
This service will handle all the settings resolution and saving for builder and map. Currently all settings are being handled independently by each of the builder components (date, residence, and accessibility) and the map component
