+++ libs/ts/movein/ngx/src/lib/modules/builder/modules/review/review.component.ts
@@ -4,7 +4,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { LocalStoreService } from '@tamu-gisc/common/ngx/local-store';
 
 import { MoveInSettings } from '../../../../interfaces/move-in-out.interface';
-import { MoveinOutService } from '../../../map/services/move-in-out.service';
+import { MoveinOutService } from '../../../map/services/move-in-out/move-in-out.service';
 
 @Component({
   selector: 'tamu-gisc-review',
