+++ libs/ts/movein/ngx/src/lib/modules/map/services/move-in-out-settings/move-in-out-settings.service.spec.ts
@@ -1,0 +1,17 @@
+import { TestBed } from '@angular/core/testing';
+
+import { MoveInOutSettingsService } from './move-in-out-settings.service';
+
+describe('MoveInOutSettingsService', () => {
+  let service: MoveInOutSettingsService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(MoveInOutSettingsService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
+
