+++ libs/ts/movein/ngx/src/lib/modules/map/services/move-in-out-settings/move-in-out-settings.service.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class MoveInOutSettingsService {}
+
