+++ libs/ts/movein/ngx/src/lib/modules/popups/move-in-out-parking-space-popup/move-in-out-parking-space-popup.component.ts
@@ -9,7 +9,7 @@ import { EsriMapService } from '@tamu-gisc/maps/esri';
 import { BaseDirectionsComponent } from '@tamu-gisc/aggiemap/ngx/popups';
 import { TripPlannerService } from '@tamu-gisc/maps/feature/trip-planner';
 
-import { MoveinOutService } from '../../map/services/move-in-out.service';
+import { MoveinOutService } from '../../map/services/move-in-out/move-in-out.service';
 
 import esri = __esri;
